Changed CallStatsLib

+++ CallStatsLib/Request/IceCandidate.cs
@@ -5,7 +5,7 @@
         public string id { get; set; }
         public string type { get; set; }
         public string ip { get; set; }
-        public int port { get; set; }
+        public long port { get; set; }
         public string candidateType { get; set; }
         public string transport { get; set; }
     }
